+++ src/index.ts
@@ -7,4 +7,5 @@ export * from "./paths.js";
 export * from "./annotations.js";
 export * from "./logger.js";
 export * from "./errors.js";
+export * from "./cache-status.js";